+++ data/src/commonMain/kotlin/regolith/data/settings/structure/SettingEntry.kt
@@ -1,0 +1,9 @@
+package regolith.data.settings.structure
+
+/**
+ * Pairs a setting with its current value.
+ */
+interface SettingEntry<T, out S: Setting<T>> {
+    val setting: S
+    val value: T
+}